Refine project setup
